Extend tests for TxList
